Create tiago_handover_objectdistorient.py
This script calculates the distance for the robot and the  handover object based on its pose. 
When the script receives the grasp pose of the handover object, it represent the object is in the grasp pose. 
However, if the object is not oriented correctly(grasp pose will not be recived  the Aruco marker cannot be detected), and the robot determines that the object is not in the grasping pose.